Greatly improve gff3 preview speed.
This is a workaround for the lack of pagination in gff3 services. It
fetches the first few ids of the paths we are interested in, and then
gets the GFF3 for a query against those ids. This is fast because the
query will be in cache in the first case, and the ids will be in cache
in the second. In both cases the result set is very small.